requirements.txt
azure-functions==1.20.0
azure-storage-queue==12.9.0
azure-data-tables==12.5.0
requests==2.32.3
pydantic==2.8.2
python-dotenv==1.0.1
anthropic>=0.39.0
codecarbon>=2.3.0
